jwt-validator-service/requirements.txt
cryptography==42.0.5
Flask==3.0.2
PyJWT==2.8.0
requests==2.31.0
waitress==3.0.0